requirements.txt
# Core dependencies
pillow>=10.0.0
numpy>=1.24.0
requests>=2.28.0

# Audio processing
sounddevice>=0.4.6
faster-whisper>=0.10.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0  # Optional but recommended for better multilingual support

# Embeddings & Retrieval
sentence-transformers>=2.2.0

# Database
# SQLite is built into Python, no additional package needed

# Optional: For GPU acceleration
# torch>=2.0.0  # If you want GPU support for embeddings/whisper
